feat: create modules complete

+++ src/modules/contacts/contacts.service.ts
@@ -3,6 +3,7 @@ import { CreateContatctDTO } from './dto/create-contact.dto';
 import { ContactsRepository } from './repositories/contact.repository';
 import { Contact } from './entitie/contact.entitie';
 import { UpdateContactDto } from './dto/update-contact.dto';
+import { Prisma } from '@prisma/client';
 
 @Injectable()
 export class ContactsService {
@@ -15,6 +16,9 @@ export class ContactsService {
 
     async findOne(id:string){
         const contact: Contact = await this.ContactsRepository.findOne(id)
+        if(!contact){
+            throw new NotFoundException("Contact not found!")
+        }
         return contact
     }
 
@@ -22,12 +26,22 @@ export class ContactsService {
         return await this.ContactsRepository.findAll()
     }
 
+
     async update(id: string, data: UpdateContactDto) {
+        const contact: Contact = await this.ContactsRepository.findOne(id)
+        if(!contact){
+            throw new NotFoundException("Contact not found!")
+        }
 
         return this.ContactsRepository.update(id, data)
     }
 
     async delete(id: string) {
+        const contact: Contact = await this.ContactsRepository.findOne(id)
+        if(!contact){
+            throw new NotFoundException("Contact not found!")
+        }
+        
         await this.ContactsRepository.delete(id)
     }
 }
